Simple page showing the geolocation of the user using Yandex Maps
There are 3 files in this repository, one HTML, one CSS and one JavaScript.
The HTML is a page easy to understand because there is not a lot of code written, I wanted to keep it simple.

The button View Location has the id of "js-location" which when will be clicked will start a function from the JavaScript file to show the geolocation of the user.

The div with the id of "map" will show the map as an image.

The CSS file is also easy to understand so I won't go into details.

The JavaScript file has 6 functions.
The first two is producing the URL required for maps to load with the latitude and longitude of the user.
I am using the second function in this example but you can change it to Google maps if you want.